Add files via upload
Name: Nishi .S. Shah
RollNo.1811045
Branch: Computer
I have selected quiz game as a topic in which first the set of instructions are displayed before playing the game. Here we have imported time module so as to perform time related operations. Then  user inputs his name and later a set of questions are displayed for which he has to enter the answer in terms of  a, b, c or d as a options . If the answer is correct then the player score gets incremented by 1 and if then answer is wrong then the correct answer is also displayed. So, in this way the user plays 10 questions and in the end his total score is displayed.